Shredding files no works properly for the kfileioslave. Ok, I lied, it doesn't, class KShred has been coded incorrectly so it doesn't just overwrite the file, it also pumps up it size.  From 1.5mb to >30mb.  That would be bad.  I can feel the pulse in my eye, so I'm not going to fix it :D.
The kioslavetest also works with it!  I'll tell you that this is a lot
better programming than KAboutData, but that doesn't say much :)  Can't I
do some string parsing? :)

Ok, enough ranting.  How many people actually read these CVS logs? Not
many.  Well, except the many people that are keeping an eye on me.  You
know who you are.  Maybe I will include a subliminal message? :)

Indenting with tabs is the one true path to nirvana.
Why am I still typing?
How many one true paths to nirvana are there?
No, really, why am I still typing.  I need to commit before my harddrive
is blown up or something.

As of this dot "." , I've written 913 characters in this cvs log.  Isn't
that pitiful.  Whoa am I bored. This is pretty lame.  Even for me, this is
an all time low.  Actually writing a CVS log that contains something.

When I was working on the KAboutdata stuff I sometimes wrote 1 word log
messages, sometimes I even used abbrev. , but this time? No!  I'm writing
a whole story!  It goes on and on!.  It in fact may never end.  If only I
could actually commit this darn thing already so that somebody could tell
me to stop writing the CVS log, but that's one of those recursive
paradoxes.

I also don't end my paragraphs properly.  I constantly create run-on
sentences and I know it, and am proud.

I'm not even on topic anymore.  Egad.  This has to end.  I must stop with
this CVS log message.

Well, how about a few jokes?  Ok, lets make fun of the name "KDE" a bit :)

"Why don't we rename the KDE Color Dialog box? We can call it KGB"
"Well, it is the Kosher Desktop Environment"
"Krap, I'm out of iDeas, Essentially" (that one was bad, I know)

The gnomers don't like KDE because it's "too much like windows".  Jerks,
the Microserfs have been copying from US!

I think this log may single-handedly make Zeus run out of disk space.

Now I feel bad.

[click]

svn path=/trunk/kdelibs/; revision=41364

+++ kio/job.cpp
@@ -22,6 +22,7 @@
 #include <config.h>
 #endif
 
+#include <iostream.h>
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <sys/stat.h>
@@ -59,7 +60,6 @@
 #include "kio/job.h"
 #include "scheduler.h"
 #include "kmimemagic.h"
-#include "kshred.h"
 
 #include "kio/renamedlg.h"
 #include "kio/skipdlg.h"
@@ -1460,7 +1460,7 @@ void DeleteJob::startNextJob()
     {
         // Stat first
         KIO::Job * job = KIO::stat( *it );
-        kDebugInfo(7007,"KIO::stat %s", (*it).url().ascii() );
+        kDebugInfo(7007,"KIO::stat (DeleteJob) %s", (*it).url().ascii() );
         state = STATE_STATING;
         addSubjob(job);
         m_srcList.remove(it);
@@ -1480,19 +1480,15 @@ void DeleteJob::deleteNextFile()
         // Use shredding ?
         if ( m_shred && (*it).isLocalFile() )
         {
-            for ( ; it != files.end() ; ++it )
-                if (!KShred::shred( (*it).path() ))
-                {
-                    m_error = ERR_CANNOT_DELETE;
-                    m_errorText = (*it).path();
-                    emit result(this);
-                    delete this;
-                    return;
-                }
-            files.clear();
-            state = STATE_DELETING_DIRS;
-            deleteNextDir();
-        } else
+            // KShred your KTie
+            KIO_ARGS << int(3) << (*it).path();
+            kDebugError("Trying to shred");
+
+            SimpleJob *job = KIO::special(KURL("file:/"), packedArgs);
+            files.remove(it);
+            addSubjob(job);
+        
+        } else 
         {
             // Normal deletion
             SimpleJob *job = KIO::file_delete( *it );
@@ -1572,9 +1568,22 @@ void DeleteJob::slotResult( Job *job )
 
                 kDebugInfo(7007," Target is a file ");
                 // Remove it
-                state = STATE_DELETING_FILES;
-                SimpleJob *job = KIO::file_delete( url );
-                addSubjob( job );
+
+                state = STATE_DELETING_FILES;                
+                if ( m_shred && url.isLocalFile() )
+                {
+                    // KShred your KTie
+                    KIO_ARGS << int(3) << url.path();
+                    kDebugError("Trying to shred");
+                    SimpleJob *job = KIO::special(KURL("file:/"), packedArgs);
+                    addSubjob(job);
+                }
+                else 
+                {
+                   // Normal deletion
+                   SimpleJob *job = KIO::file_delete(url);
+                   addSubjob( job );
+                }
             }
         }
         break;
